+++ src/lermitage/intellij/extra/icons/ModelType.java
@@ -1,0 +1,6 @@
+package lermitage.intellij.extra.icons;
+
+public enum ModelType {
+    FILE,
+    DIR
+}